Refactor questionnaire service structure and implement initial database models and settings
