--- .dockerignore
@@ -1,42 +1,0 @@
-# Database files
-database/
-*.db
-*.sqlite
-*.sqlite3
-
-# Environment and configuration
-.env
-.env.*
-*.env
-venv/
-.venv/
-env/
-.python-version
-
-# Cache and temporary files
-__pycache__/
-*.py[cod]
-*$py.class
-.pytest_cache/
-.coverage
-htmlcov/
-.tox/
-.nox/
-
-# Development tools
-.git/
-.github/
-.idea/
-.vscode/
-*.swp
-*.swo
-
-# Logs
-logs/
-*.log
-npm-debug.log*
-
-# Build artifacts
-build/
-dist/
-*.egg-info/
