--- .env.example
@@ -1,13 +1,0 @@
-# Database configuration
-DB_USER=questionnaire_user
-DB_PASSWORD=questionnaire_password
-DB_HOST=mysql
-DB_PORT=3306
-DB_NAME=questionnaire_db
-MYSQL_ROOT_PASSWORD=rootpassword
-
-# Application configuration
-APP_SECRET_KEY=234erfdftyewerwqasda
-
-# Domain settings
-DOMAIN=trainings.admin.asuta.or.tz
