+++ alembic.ini
@@ -9,7 +9,7 @@ script_location = alembic
 # Uncomment the line below if you want the files to be prepended with date and time
 # see https://alembic.sqlalchemy.org/en/latest/tutorial.html#editing-the-ini-file
 # for all available tokens
-file_template = %%(year)d%%(month).2d%%(day).2d_%%(hour).2d%%(minute).2d%%(second).2d_%%(rev)s_%%(slug)s
+# file_template = %%(year)d_%%(month).2d_%%(day).2d_%%(hour).2d%%(minute).2d-%%(rev)s_%%(slug)s
 
 # sys.path path, will be prepended to sys.path if present.
 # defaults to the current working directory.
@@ -21,19 +21,19 @@ prepend_sys_path = .
 # Any required deps can installed by adding `alembic[tz]` to the pip requirements
 # string value is passed to ZoneInfo()
 # leave blank for localtime
-timezone = UTC
+# timezone =
 
 # max length of characters to apply to the "slug" field
-truncate_slug_length = 40
+# truncate_slug_length = 40
 
 # set to 'true' to run the environment during
 # the 'revision' command, regardless of autogenerate
-revision_environment = false
+# revision_environment = false
 
 # set to 'true' to allow .pyc and .pyo files without
 # a source .py file to be detected as revisions in the
 # versions/ directory
-sourceless = false
+# sourceless = false
 
 # version location specification; This defaults
 # to alembic/versions.  When using multiple version
@@ -63,9 +63,12 @@ version_path_separator = os
 # are written from script.py.mako
 # output_encoding = utf-8
 
-# Change this line to point to your SQLite database
-sqlalchemy.url = postgresql://questionnaire:questionnaire@db/questionnaire
+# Update the SQLAlchemy URL to use the correct format:
+sqlalchemy.url = sqlite:///./questionnaire.db
+# Or if you're using PostgreSQL: postgresql://username:password@localhost/dbname
+# Or if you're using MySQL: mysql://username:password@localhost/dbname
 
+# Make sure there are no quotes around the URL
 
 [post_write_hooks]
 # post_write_hooks defines scripts or Python functions that are run
