+++ alembic/env.py
@@ -1,45 +1,43 @@
-from __future__ import with_statement
-from alembic import context
-from sqlalchemy import engine_from_config, pool
 from logging.config import fileConfig
+
+from sqlalchemy import engine_from_config
+from sqlalchemy import pool
+
+from alembic import context
+
 import os
 import sys
 
-# Add the project root to the Python path
-sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+# Add the parent directory to sys.path
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+# Import the SQLAlchemy components from your app
+from database import Base, SQLALCHEMY_DATABASE_URL
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
 config = context.config
 
-# Import your database URL from your project
-try:
-    from database import DATABASE_URL
-    config.set_main_option('sqlalchemy.url', DATABASE_URL)
-except ImportError:
-    # Fall back to the URL in alembic.ini if import fails
-    pass
-
 # Interpret the config file for Python logging.
-fileConfig(config.config_file_name)
+# This line sets up loggers basically.
+if config.config_file_name is not None:
+    fileConfig(config.config_file_name)
 
-# Import your models
-# Adjust the import path to match your project structure
-try:
-    from src.models import Base
-except ImportError:
-    try:
-        from models import Base
-    except ImportError:
-        raise Exception(
-            "Could not import Base from models. "
-            "Make sure your models are properly defined with a Base object."
-        )
-
+# add your model's MetaData object here
+# for 'autogenerate' support
 target_metadata = Base.metadata
 
-def run_migrations_offline():
-    """Run migrations in 'offline' mode."""
+def run_migrations_offline() -> None:
+    """Run migrations in 'offline' mode.
+
+    This configures the context with just a URL
+    and not an Engine, though an Engine is acceptable
+    here as well.  By skipping the Engine creation
+    we don't even need a DBAPI to be available.
+
+    Calls to context.execute() here emit the given string to the
+    script output.
+    """
     url = config.get_main_option("sqlalchemy.url")
     context.configure(
         url=url,
@@ -52,17 +50,25 @@ def run_migrations_offline():
         context.run_migrations()
 
 
-def run_migrations_online():
-    """Run migrations in 'online' mode."""
+def run_migrations_online() -> None:
+    """Run migrations in 'online' mode.
+
+    In this scenario we need to create an Engine
+    and associate a connection with the context.
+    """
+    # Use the actual URL from your application
+    configuration = config.get_section(config.config_ini_section)
+    configuration['sqlalchemy.url'] = SQLALCHEMY_DATABASE_URL
+    
     connectable = engine_from_config(
-        config.get_section(config.config_ini_section),
+        configuration,
         prefix="sqlalchemy.",
         poolclass=pool.NullPool,
     )
 
     with connectable.connect() as connection:
         context.configure(
-            connection=connection,
+            connection=connection, 
             target_metadata=target_metadata
         )
