--- alembic/versions/a6481145521a_initial_migration.py
@@ -1,52 +1,0 @@
-"""Initial migration
-
-Revision ID: a6481145521a
-Revises: 
-Create Date: 2025-03-19 12:58:20.068679
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'a6481145521a'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('form_submissions',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('(CURRENT_TIMESTAMP)'), nullable=True),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=True),
-    sa.Column('jina_kamili', sa.String(length=255), nullable=True),
-    sa.Column('jinsi', sa.String(length=50), nullable=True),
-    sa.Column('umri', sa.String(length=50), nullable=True),
-    sa.Column('barua_pepe', sa.String(length=255), nullable=True),
-    sa.Column('nambari_simu', sa.String(length=255), nullable=True),
-    sa.Column('ngazi_elimi', sa.String(length=255), nullable=True),
-    sa.Column('jina_kozi', sa.String(length=255), nullable=True),
-    sa.Column('mwaka_kuhitimu', sa.String(length=50), nullable=True),
-    sa.Column('taasisi_mwisho', sa.String(length=255), nullable=True),
-    sa.Column('ujuzi_kompyuta', sa.String(length=50), nullable=True),
-    sa.Column('lugha', sa.Text(), nullable=True),
-    sa.Column('msaada', sa.Text(), nullable=True),
-    sa.Column('pasipoti_path', sa.String(length=255), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_form_submissions_id'), 'form_submissions', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_form_submissions_id'), table_name='form_submissions')
-    op.drop_table('form_submissions')
-    # ### end Alembic commands ###
